style: rename file to maintain consistency
