+++ libs/lib-server/core/src/lib/connection-settings/index.ts
@@ -1,1 +1,1 @@
-export * from './transformConnectionTypeSettings';
+export * from './transform-connection-type-settings';
